+++ CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [PEP 440](https://www.python.org/dev/peps/pep-0440/)
 and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0]
+
+### Changed
+* `create_browse` now takes VV.tif and VH.tif paths as input rather than a granule name
+
+### Removed
+* `create_browse` option to download a granule via CMR
+* `create_browse` option to upload the output rgb to an S3 bucket
+
 ## [0.3.0]
 
 ### Changed